+++ Code/crop_image.h
@@ -7,5 +7,5 @@
 #include "miscelanious_utils.h"
 
 void crop(image * img, selected_area * selected, status * img_status);
-int exe_crop(uint_fast8_t ***matrix, selected_area *selected, int height);
+int exe_crop(double ***matrix, selected_area *selected, int height);
 int exe_crop_bw(uint_fast8_t ***matrix, selected_area *selected, int height);
